requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.3.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.2
joblib==1.3.2
pydantic==2.5.2